extras/previous_versions/v2/trigger_script/requirements_trigger.txt
python-dotenv>=1.0.1


